```
AFTER

Benchmark                                                (groupByTypes)  (operatorType)  (rowsPerGroup)  Mode  Cnt    Score    Error  Units
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash               1  avgt   30  279,836 ±  4,167  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash              10  avgt   30   62,129 ±  0,370  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash            1000  avgt   30   37,873 ±  1,011  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash               1  avgt   30  375,288 ± 12,203  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash              10  avgt   30   85,357 ±  0,651  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash            1000  avgt   30   57,710 ±  2,811  ms/op

BEFORE:

Benchmark                                                (groupByTypes)  (operatorType)  (rowsPerGroup)  Mode  Cnt    Score    Error  Units
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash               1  avgt   30  363,063 ±  5,997  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash              10  avgt   30   68,201 ±  0,587  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark         varchar            hash            1000  avgt   30   41,893 ±  0,568  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash               1  avgt   30  466,960 ± 13,922  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash              10  avgt   30   91,118 ±  1,326  ms/op
BenchmarkHashAndStreamingAggregationOperators.benchmark           mixed            hash            1000  avgt   30   58,448 ±  0,371  ms/op
```